+++ testgen/src/consensus.rs
@@ -12,7 +12,7 @@ pub fn default_consensus_params() -> consensus::Params {
         evidence: evidence::Params {
             max_age_num_blocks: 100000,
             max_age_duration: evidence::Duration(std::time::Duration::new(48 * 3600, 0)),
-            max_num: 0,
+            max_bytes: 1048576,
         },
         validator: consensus::params::ValidatorParams {
             pub_key_types: vec![Algorithm::Ed25519],